Expand documentation
"Oh rewrite it in Rust it'll be fun and you'll get to have a better type
system and it'll be faster there's no reason not to" 1200 lines and 2.5
days later and I've just finished StageType, probably the least
complicated part of it lmao. Anyway now I can move on to more headaches!