Change home directory library crate
`shell-expand` has a whole tree of dependencies which ends at
`windows-sys` 0.48.0, which takes at least 3 minutes to compile the docs
for, which is completely unacceptable. `home` uses 0.52.0 which realises
how long its docs take to compile and hide it behind a flag.

Serde also takes a minute but that's nothing compared to `windows-sys`.