Add webpage opener.
JFC all of these libraries are complicated. What happened to JavaScript
where it's just `fetch` and that's it and you didn't need main to be
async if you wanted the tiniest function to be async. And tokio's docs
take years to compile and like... I don't need that. I'm reading 7
files. Anyway `ureq` has a load of dependencies but they're really small
so it's fine, thank goodness.