Add prototype encounters.
Significant changes will need to be made for it to actually work on any
units, and then significant changes will need to be made for it to just
work on any page, and even then there will need to be work done to get
it to deal with edge cases.

Also really needs refactoring.